moves unsafeHTML and spreadAttributes to other directives
